Refactor app_config and add MemberSection and TeamSection classes for improved configuration management
